Upgraded migrations to fix a few small problems

+++ askbot/migrations/0092_postize_vote_and_activity.py
@@ -7,16 +7,8 @@ from django.db import models
 class Migration(DataMigration):
 
     def forwards(self, orm):
-        if orm.Activity.objects.exists():
-            try:
-                ct_post = orm['contenttypes.ContentType'].objects.get(app_label='askbot', model='post')
-            except orm['contenttypes.ContentType'].DoesNotExist:
-                print "The new `Post` models doesn't yet have a matching ContentType instance,"
-                print "but it'll be created by Django automatically, so just re-run ./manage.py migrate"
-                raise ValueError('Just re-run ./manage.py migrate')
 
-        ###
-
+        # TODO: Speed this up by prefetching all votes ?
         for v in orm.Vote.objects.all():
             if (v.content_type.app_label, v.content_type.model) == ('askbot', 'question'):
                 v.voted_post = orm.Post.objects.get(self_question__id=v.object_id)
@@ -30,42 +22,26 @@ class Migration(DataMigration):
 
         ###
 
-        if not orm.Activity.objects.exists():
-            return
+        # ContentType for Post model might not yet be present in the database
+        # (if migrations are applied in a row then contenttypes update is not called between them)
+        ct_post, c = orm['contenttypes.ContentType'].objects.get_or_create(app_label='askbot', model='post', defaults={'name': 'post'})
+
+        print "!!! [Red] Remember to not remove the old content types for Question, Answer and Comment models until all migrations succeed!"
 
         abandoned_activities = []
 
         for a in orm.Activity.objects.all():
+            # test if content_object for this activity exists - there might be a bunch of "abandoned" activities
+            model_signature = '.'.join([a.content_type.app_label, a.content_type.model])
+            if not orm[model_signature].objects.filter(id=a.object_id).exists():
+                abandoned_activities.append(a)
+                continue
+
             save = False
-            content_tuple = (a.content_type.app_label, a.content_type.model)
 
-            if content_tuple == ('askbot', 'question'):
-                try:
-                    orm.Question.objects.get(id=a.object_id)
-                except orm.Question.DoesNotExist:
-                    # TODO: Maybe question can be recovered from activity.question denormalized field
-                    #      only only if activity.question is broken treat this activity as abandoned ?
-                    abandoned_activities.append(a.id)
-                    continue
-                a.object_id = orm.Post.objects.get(self_question__id=a.object_id).id
-                save = True
-
-            elif content_tuple == ('askbot', 'answer'):
-                try:
-                    orm.Answer.objects.get(id=a.object_id)
-                except orm.Answer.DoesNotExist:
-                    abandoned_activities.append(a.id)
-                    continue
-                a.object_id = orm.Post.objects.get(self_answer__id=a.object_id).id
-                save = True
-
-            elif content_tuple == ('askbot', 'comment'):
-                try:
-                    orm.Comment.objects.get(id=a.object_id)
-                except orm.Comment.DoesNotExist:
-                    abandoned_activities.append(a.id)
-                    continue
-                a.object_id = orm.Post.objects.get(self_comment__id=a.object_id).id
+            model = a.content_type.model
+            if a.content_type.app_label == 'askbot' and model in ('question', 'answer', 'comment'):
+                a.object_id = orm.Post.objects.get(**{'self_%s__id' % model: a.object_id}).id
                 save = True
 
             if a.question:
@@ -76,10 +52,15 @@ class Migration(DataMigration):
                 a.content_type = ct_post
                 a.save()
 
-        # INFO: non-batch delete to process all dependencies, if there are any
-        print "!!! Abandoned activities (num=%d):" % len(abandoned_activities), abandoned_activities
-        for a in orm.Activity.objects.filter(id__in=abandoned_activities):
-            a.delete()
+        if abandoned_activities:
+            # Remove "abandoned" activities
+            abandoned_activities_lst = [
+                (a.id, '.'.join([a.content_type.app_label, a.content_type.model]), a.object_id)
+                for a in abandoned_activities
+            ]
+            print "!!! Abandoned activities (num=%d):" % len(abandoned_activities), abandoned_activities_lst
+            for a in abandoned_activities:
+                a.delete()
 
 
     def backwards(self, orm):
